домашка 19  6 задача
