--- Lesson16/src/main/java/Task2/model/Animals.java
@@ -1,26 +1,0 @@
-package Task2.model;
-
-import java.io.Serializable;
-
-public class Animals implements AnimalsAware, Serializable {
-    private String name;
-    private int age;
-
-    public Animals(String name, int age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    @Override
-    public void say() {
-        System.out.println("Говорю!");
-    }
-
-    @Override
-    public String toString() {
-        return "Animals{" +
-                "name='" + name + '\'' +
-                ", age=" + age +
-                '}';
-    }
-}
