--- Lesson16/src/main/java/Task2/model/AnimalsAware.java
@@ -1,5 +1,0 @@
-package Task2.model;
-
-public interface AnimalsAware {
-    void say();
-}
