--- Lesson16/src/main/java/Task4/Task4.java
@@ -1,13 +1,0 @@
-package Task4;
-
-import static Task4.Calculator.*;
-
-public class Task4 {
-    public static void main(String[] args) {
-        System.out.println(multiply(2.341, 2456));
-        System.out.println(subtraction(4563, 2.22));
-        System.out.println(sum(2.3232123, 2.2351));
-        System.out.println(divide(6, 2));
-        System.out.println(divide(3, 2));
-    }
-}
