+++ Lesson16/src/main/java/task2/model/AnimalsAware.java
@@ -1,0 +1,5 @@
+package task2.model;
+
+public interface AnimalsAware {
+    void say();
+}
