+++ Lesson16/src/main/java/task3/model/MinMax.java
@@ -1,0 +1,29 @@
+package task3.model;
+
+import lombok.Getter;
+import lombok.NonNull;
+
+import java.util.Arrays;
+
+@Getter
+public class MinMax<T extends Number> {
+    private final T[] arrayOfNumbers;
+
+    public MinMax(@NonNull T[] array) {
+//        Objects.requireNonNull(array);
+        if (array.length == 0) {
+            throw new IllegalArgumentException("array is empty");
+        }
+        this.arrayOfNumbers = array;
+    }
+
+    public T getMax() {
+        Arrays.sort(arrayOfNumbers);
+        return arrayOfNumbers[arrayOfNumbers.length - 1];
+    }
+
+    public T getMin() {
+        Arrays.sort(arrayOfNumbers);
+        return arrayOfNumbers[0];
+    }
+}
