--- Lesson19/src/main/java/Task4/model/Car.java
@@ -1,11 +1,0 @@
-package Task4.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Car {
-    private String num;
-    private int year;
-}
