--- Lesson19/src/main/java/Task5/model/Person.java
@@ -1,11 +1,0 @@
-package Task5.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Person {
-    private String firstName;
-    private String secondName;
-}
