--- Lesson19/src/main/java/Task6/model/Book.java
@@ -1,12 +1,0 @@
-package Task6.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Book {
-    String author;
-    String title;
-    int theYearOfPublishing;
-}
