--- Lesson19/src/main/java/Task6/model/EmailAddress.java
@@ -1,11 +1,0 @@
-package Task6.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class EmailAddress {
-    String emailAddress;
-    String additionalInformation;
-}
