--- Lesson19/src/main/java/Task6/model/Library.java
@@ -1,13 +1,0 @@
-package Task6.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-public class Library {
-    List<Book> books;
-    List<Reader> readers;
-}
