--- Lesson19/src/main/java/Task6/model/Reader.java
@@ -1,17 +1,0 @@
-package Task6.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-public class Reader {
-    String fullName;
-    EmailAddress eMail;
-    boolean agreement;
-    List<Book> readersBooks;
-
-
-}
