+++ Lesson19/src/main/java/task4/model/Car.java
@@ -1,0 +1,11 @@
+package task4.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Car {
+    private String num;
+    private int year;
+}
