+++ Lesson19/src/main/java/task5/model/Person.java
@@ -1,0 +1,11 @@
+package task5.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Person {
+    private String firstName;
+    private String secondName;
+}
