+++ Lesson19/src/main/java/task6/models/EmailAddress.java
@@ -1,0 +1,16 @@
+package task6.models;
+
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class EmailAddress {
+    private String email;
+    private String additionalInformation;
+
+    public EmailAddress(String email) {
+        this.email = email;
+    }
+}
