+++ Lesson19/src/main/java/task6/models/Library.java
@@ -1,0 +1,30 @@
+package task6.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Library {
+    private List<Book> books;
+    private List<Reader> readers;
+
+    public List<Book> getBooks() {
+        if (books == null) {
+            books = new ArrayList<>();
+        }
+        return books;
+    }
+
+    public List<Reader> getReaders() {
+        if (readers == null) {
+            readers = new ArrayList<>();
+        }
+        return readers;
+    }
+}
